+++ src/pages/Pool/PoolManagement.tsx
@@ -799,7 +799,14 @@ export default function PoolManagement({
   const pairPoolDepositFilter = usePoolDepositFilterForPair(
     tokenA && tokenB ? [tokenA, tokenB] : ['', '']
   );
-  const userPositionsContext = useUserPositionsContext(pairPoolDepositFilter);
+  const userUnstakedContext = useUserPositionsContext(pairPoolDepositFilter);
+  const userStakedContext = useUserPositionsContext(
+    pairPoolDepositFilter,
+    true
+  );
+  const userPositionsContext = useMemo(() => {
+    return [...userUnstakedContext, ...userStakedContext];
+  }, [userStakedContext, userUnstakedContext]);
 
   const [{ isValidating: isValidatingEdit }, sendEditRequest] =
     useEditLiquidity();
@@ -1627,12 +1634,14 @@ export default function PoolManagement({
                 setEditedUserPosition={setEditedUserPosition}
                 viewableMinIndex={viewableMinIndex}
                 viewableMaxIndex={viewableMaxIndex}
+                edgePriceIndex={edgePriceIndex}
               />
             ) : (
               <MyNewPositionTableCard
                 tokenA={tokenA}
                 tokenB={tokenB}
                 userTicks={userTicks}
+                edgePriceIndex={edgePriceIndex}
               />
             )}
           </div>